update PLCL/R and include spline

+++ src/prediction.h
@@ -20,7 +20,7 @@ void print_curr_trajectories();
 void print_trajectories(const map<int, deque<Vehicle> > &prediction_traj);
 void update_trajectories(const vector<vector<int>> &sensorfusion);
 
-map<int, deque<Vehicle>> do_prediction(int dt_s = 1);
+map<int, deque<Vehicle>> do_prediction(double dt_s = 1.);
 unsigned int num_vehicles_;
 
 // Map of the car id and its trajectories
